qualitair/pyproject.toml
[tool.poetry]
name = "qualitair"
version = "0.0.0"
description = "Air quality monitoring"
authors = ["Benj Fassbind <randombenj@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pimoroni-sgp30 = "^0.0.2"
aiohttp = "^3.7.3"
smbus2 = "^0.4.1"
tortoise-orm = "^0.16.20"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
